adding groups is now complete, needs adjustment to the notificaiton of image loading success.
i also need to add the list of members of the group (so that the members can be displayed).

updated constraints

updating the ability the set location and return the address through the means of intents

need to actually update to the real time database the JSON file that is the groups, however this needs to be done in such a way that groups are maintained and mutable therefore requires being researched.