updating groups fixed.
ability for database to track which users are members and requested members of which group implemented.
this allows for easier methods of getting current users groups with minimal requests.

however this is not fully tested and requires validation tomorrow.

new feature to display current groups will now be added, with a max limit of 10 on the query for now, this is due to firebase restrictions, a work around can be made however awaiting approval from lecturer