search feature fully implemented lots of changes around the adapter classes needs its own adapter for searchgroups.
need to refresh the group list upon document changes.

need or need to implement a way for the group overview to spot when a group has been deleted and display a message, to reduce search cost

current groups has an error of displaying a blank cardview at the start when overpopulated, need to check if this is because of a database error (might've accidentally duplicated memberships when manually editing the database) or if the holder is being duplicated in client code.