loading bars added to login pages, and the ability to see the group members on the map has been added.
next we need to add the custom info view holders to the map overview and update the information on the map periodically, this should be controlled by clientside as updating every time something changes about the group is expensive (we could write exceptions for this such as the meeting location changing and so fourth.

loading screens for adding groups need to be added next, and confirm popup windows.

near completion.