Complete refactor of how groups are uploaded, changed from 22k daily writes for a medicore user to 1k daily writes (at a push) added info windows however images do not load intime to update and this needs to be figured out.
also need to figure out how to display users potentially and their eta which would have to update (maybe look for alternatives for this, as the eta could be a hint on the route?)

this is because user cards cannot update easily once being generated because it generates an image.

everything is tested, should retain the same functionality as beforehand.