ability for a group to be completed, ability for the group members to register their arrival has been added.
all that is left is adding settings page and giving the map the ability to update.

updating should only happen on the general overview, once the group is inpsected it should freeze updating that group because then it could jut the group around while the user is looking at it, which is not ideal; instead maybe push a notification the group has updated; the users pos should update however.

settings page should be basic, should allow the user to select what color they'd want and update their profile, maybe also implement how long between updates of the mapoverview.